decomposed processor class into a separate pipeline class, and a handlercontext which is passed into handler functions. better separation of concerns. circular symbol important of HandlerContext and ProcessorInterface resolved with forward references
