--- src/koi_net/processor/__init__.py
@@ -1,1 +1,0 @@
-from .interface import ProcessorInterface
