+++ src/koi_net/processor/default_handlers.py
@@ -4,8 +4,8 @@ import logging
 from rid_lib.ext.bundle import Bundle
 from rid_lib.types import KoiNetNode, KoiNetEdge
 from koi_net.protocol.node import NodeType
-from .interface import ProcessorInterface
 from .handler import KnowledgeHandler, HandlerType, STOP_CHAIN
+from .handler_context import HandlerContext
 from .knowledge_object import KnowledgeObject, KnowledgeSource
 from ..protocol.event import Event, EventType
 from ..protocol.edge import EdgeProfile, EdgeStatus, EdgeType
@@ -17,12 +17,12 @@ logger = logging.getLogger(__name__)
 # RID handlers
 
 @KnowledgeHandler.create(HandlerType.RID)
-def basic_rid_handler(processor: ProcessorInterface, kobj: KnowledgeObject):
+def basic_rid_handler(ctx: HandlerContext, kobj: KnowledgeObject):
     """Default RID handler.
     
     Blocks external events about this node. Allows `FORGET` events if RID is known to this node.
     """
-    if (kobj.rid == processor.identity.rid and 
+    if (kobj.rid == ctx.identity.rid and 
         kobj.source == KnowledgeSource.External):
         logger.debug("Don't let anyone else tell me who I am!")
         return STOP_CHAIN
@@ -34,12 +34,13 @@ def basic_rid_handler(processor: ProcessorInterface, kobj: KnowledgeObject):
 # Manifest handlers
 
 @KnowledgeHandler.create(HandlerType.Manifest)
-def basic_manifest_handler(processor: ProcessorInterface, kobj: KnowledgeObject):
+def basic_manifest_handler(ctx: HandlerContext, kobj: KnowledgeObject):
     """Default manifest handler.
     
     Blocks manifests with the same hash, or aren't newer than the cached version. Sets the normalized event type to `NEW` or `UPDATE` depending on whether the RID was previously known to this node.
     """
-    prev_bundle = processor.cache.read(kobj.rid)
+    # NOTE: can be replaced by deref
+    prev_bundle = ctx.cache.read(kobj.rid)
 
     if prev_bundle:
         if kobj.manifest.sha256_hash == prev_bundle.manifest.sha256_hash:
@@ -66,7 +67,7 @@ def basic_manifest_handler(processor: ProcessorInterface, kobj: KnowledgeObject)
     rid_types=[KoiNetEdge], 
     source=KnowledgeSource.External,
     event_types=[EventType.NEW, EventType.UPDATE])
-def edge_negotiation_handler(processor: ProcessorInterface, kobj: KnowledgeObject):
+def edge_negotiation_handler(ctx: HandlerContext, kobj: KnowledgeObject):
     """Handles basic edge negotiation process.
     
     Automatically approves proposed edges if they request RID types this node can provide (or KOI nodes/edges). Validates the edge type is allowed for the node type (partial nodes cannot use webhooks). If edge is invalid, a `FORGET` event is sent to the other node.
@@ -75,14 +76,14 @@ def edge_negotiation_handler(processor: ProcessorInterface, kobj: KnowledgeObjec
     edge_profile = kobj.bundle.validate_contents(EdgeProfile)
 
     # indicates peer subscribing to me
-    if edge_profile.source == processor.identity.rid:     
+    if edge_profile.source == ctx.identity.rid:     
         if edge_profile.status != EdgeStatus.PROPOSED:
             return
         
         logger.debug("Handling edge negotiation")
         
         peer_rid = edge_profile.target
-        peer_bundle = processor.network.cache.read(peer_rid)
+        peer_bundle = ctx.cache.read(peer_rid)
         
         if not peer_bundle:
             logger.warning(f"Peer {peer_rid} unknown to me")
@@ -92,7 +93,7 @@ def edge_negotiation_handler(processor: ProcessorInterface, kobj: KnowledgeObjec
         
         # explicitly provided event RID types and (self) node + edge objects
         provided_events = (
-            *processor.identity.profile.provides.event,
+            *ctx.identity.profile.provides.event,
             KoiNetNode, KoiNetEdge
         )
         
@@ -109,7 +110,7 @@ def edge_negotiation_handler(processor: ProcessorInterface, kobj: KnowledgeObjec
         
         if abort:
             event = Event.from_rid(EventType.FORGET, kobj.rid)
-            processor.network.push_event_to(event, peer_rid, flush=True)
+            ctx.network.push_event_to(event, peer_rid, flush=True)
             return STOP_CHAIN
 
         else:
@@ -118,10 +119,10 @@ def edge_negotiation_handler(processor: ProcessorInterface, kobj: KnowledgeObjec
             edge_profile.status = EdgeStatus.APPROVED
             updated_bundle = Bundle.generate(kobj.rid, edge_profile.model_dump())
       
-            processor.handle(bundle=updated_bundle, event_type=EventType.UPDATE)
+            ctx.handle(bundle=updated_bundle, event_type=EventType.UPDATE)
             return
               
-    elif edge_profile.target == processor.identity.rid:
+    elif edge_profile.target == ctx.identity.rid:
         if edge_profile.status == EdgeStatus.APPROVED:
             logger.debug("Edge approved by other node!")
 
@@ -129,7 +130,7 @@ def edge_negotiation_handler(processor: ProcessorInterface, kobj: KnowledgeObjec
 # Network handlers
 
 @KnowledgeHandler.create(HandlerType.Network, rid_types=[KoiNetNode])
-def coordinator_contact(processor: ProcessorInterface, kobj: KnowledgeObject):
+def coordinator_contact(ctx: HandlerContext, kobj: KnowledgeObject):
     node_profile = kobj.bundle.validate_contents(NodeProfile)
             
     # looking for event provider of nodes
@@ -137,54 +138,54 @@ def coordinator_contact(processor: ProcessorInterface, kobj: KnowledgeObject):
         return
     
     # prevents coordinators from attempting to form a self loop
-    if kobj.rid == processor.identity.rid:
+    if kobj.rid == ctx.identity.rid:
         return
     
     # already have an edge established
-    if processor.network.graph.get_edge(
+    if ctx.graph.get_edge(
         source=kobj.rid,
-        target=processor.identity.rid,
+        target=ctx.identity.rid,
     ) is not None:
         return
     
     logger.info("Identified a coordinator!")
     logger.info("Proposing new edge")
     
-    if processor.identity.profile.node_type == NodeType.FULL:
+    if ctx.identity.profile.node_type == NodeType.FULL:
         edge_type = EdgeType.WEBHOOK
     else:
         edge_type = EdgeType.POLL
     
     # queued for processing
-    processor.handle(bundle=generate_edge_bundle(
+    ctx.handle(bundle=generate_edge_bundle(
         source=kobj.rid,
-        target=processor.identity.rid,
+        target=ctx.identity.rid,
         edge_type=edge_type,
         rid_types=[KoiNetNode]
     ))
     
     logger.info("Catching up on network state")
     
-    payload = processor.network.request_handler.fetch_rids(
+    payload = ctx.network.request_handler.fetch_rids(
         node=kobj.rid, 
         rid_types=[KoiNetNode]
     )
     for rid in payload.rids:
-        if rid == processor.identity.rid:
+        if rid == ctx.identity.rid:
             logger.info("Skipping myself")
             continue
-        if processor.cache.exists(rid):
+        if ctx.cache.exists(rid):
             logger.info(f"Skipping known RID '{rid}'")
             continue
         
         # marked as external since we are handling RIDs from another node
         # will fetch remotely instead of checking local cache
-        processor.handle(rid=rid, source=KnowledgeSource.External)
+        ctx.handle(rid=rid, source=KnowledgeSource.External)
     logger.info("Done")
     
 
 @KnowledgeHandler.create(HandlerType.Network)
-def basic_network_output_filter(processor: ProcessorInterface, kobj: KnowledgeObject):
+def basic_network_output_filter(ctx: HandlerContext, kobj: KnowledgeObject):
     """Default network handler.
     
     Allows broadcasting of all RID types this node is an event provider for (set in node profile), and other nodes have subscribed to. All nodes will also broadcast about their own (internally sourced) KOI node, and KOI edges that they are part of, regardless of their node profile configuration. Finally, nodes will also broadcast about edges to the other node involved (regardless of if they are subscribed)."""
@@ -192,25 +193,25 @@ def basic_network_output_filter(processor: ProcessorInterface, kobj: KnowledgeOb
     involves_me = False
     if kobj.source == KnowledgeSource.Internal:
         if (type(kobj.rid) == KoiNetNode):
-            if (kobj.rid == processor.identity.rid):
+            if (kobj.rid == ctx.identity.rid):
                 involves_me = True
         
         elif type(kobj.rid) == KoiNetEdge:
             edge_profile = kobj.bundle.validate_contents(EdgeProfile)
             
-            if edge_profile.source == processor.identity.rid:
+            if edge_profile.source == ctx.identity.rid:
                 logger.debug(f"Adding edge target '{edge_profile.target!r}' to network targets")
                 kobj.network_targets.update([edge_profile.target])
                 involves_me = True
                 
-            elif edge_profile.target == processor.identity.rid:
+            elif edge_profile.target == ctx.identity.rid:
                 logger.debug(f"Adding edge source '{edge_profile.source!r}' to network targets")
                 kobj.network_targets.update([edge_profile.source])
                 involves_me = True
     
-    if (type(kobj.rid) in processor.identity.profile.provides.event or involves_me):
+    if (type(kobj.rid) in ctx.identity.profile.provides.event or involves_me):
         # broadcasts to subscribers if I'm an event provider of this RID type OR it involves me
-        subscribers = processor.network.graph.get_neighbors(
+        subscribers = ctx.graph.get_neighbors(
             direction="out",
             allowed_type=type(kobj.rid)
         )
